flask version/ticket-platform-backend/requirements.txt
blinker==1.9.0
cachelib==0.13.0
chardet==5.2.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
Flask-Mail==0.10.0
Flask-Session==0.8.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
pillow==11.2.1
PyJWT==2.10.1
qrcode==8.1
reportlab==4.4.0
Werkzeug==3.1.3
